hellodjango/requirements.txt
Django==1.11.7
psycopg2==2.7.3.2
django-redis==4.4.4
